// src/enums/store.enum.ts
export enum dashboard {
  DASHBOARD = 'dashboard',
  CAPACITY = 'capacity',
  CHARECTER = 'character',
  CAPITAL = 'capital',
}
export enum onboardFlow {
  CAPACITY = 'capacity',
  CHARECTER = 'character',
  CAPITAL = 'capital',
  COMPLETED = 'completed',
  KBA = 'kba',
}
export enum forms {
  PHONE = 'phone',
  SSN = 'ssn',
}
export enum onboardStep {
  WELCOME = 'welcome',
  FORM = 'form',
  COMPLETE = 'complete',
}